pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nietzsche"
version = "0.1.2"
description = "A CLI tool for generating PowerPoint presentations"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Yann Debray"}
]
dependencies = [
    "python-pptx>=1.0.2",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
power = "power.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
